# levenshtein: Zig: use memcopy and avoid using allocators