The Levenshtein benchmark graphics made me suspicious something was up with the Zig implementation of the new reference. My understanding of Zig performance is that it is generally at C level or faster. Indeed, by avoiding try/catch allocators and using memcopy instead of assigning via a swap variable I could get 5X faster and almost where C is. I think that someone who knows Zig better than I do (this is the first Zig I've ever touched) can find more performance.

## Before

```
Benchmarking Zig
Benchmark 1:  ./zig/code aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa...
  Time (mean ± σ):      50.8 ms ±  12.6 ms    [User: 10.2 ms, System: 40.2 ms]
  Range (min … max):    37.5 ms …  73.3 ms    7 runs
 

Benchmarking C
Benchmark 1:  ./c/code aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa...
  Time (mean ± σ):       8.6 ms ±   1.0 ms    [User: 6.8 ms, System: 1.4 ms]
  Range (min … max):     7.7 ms …  10.7 ms    7 runs
 

Benchmarking Rust
Benchmark 1:  ./rust/target/release/code aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa...
  Time (mean ± σ):      11.7 ms ±   0.3 ms    [User: 10.2 ms, System: 1.1 ms]
  Range (min … max):    11.2 ms …  11.9 ms    7 runs
```

## After

```
Benchmarking Zig
Benchmark 1:  ./zig/code aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa...
  Time (mean ± σ):       9.6 ms ±   0.2 ms    [User: 8.3 ms, System: 1.0 ms]
  Range (min … max):     9.4 ms …  10.1 ms    7 runs
 

Benchmarking C
Benchmark 1:  ./c/code aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa...
  Time (mean ± σ):       8.0 ms ±   0.3 ms    [User: 6.6 ms, System: 1.1 ms]
  Range (min … max):     7.7 ms …   8.7 ms    7 runs
 

Benchmarking Rust
Benchmark 1:  ./rust/target/release/code aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa...
  Time (mean ± σ):      11.4 ms ±   0.3 ms    [User: 10.1 ms, System: 1.0 ms]
  Range (min … max):    11.0 ms …  11.8 ms    7 runs
```